+++ index.js
@@ -1,5 +1,5 @@
 import { AppRegistry } from 'react-native';
-import App from './page/app/App';
+import App from './page/App';
 import { name as appName } from './app.json';
 
 // app
