+++ package.json
@@ -14,11 +14,15 @@
     "lint:fix": "eslint --fix './**/*.{ts,tsx}'"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.2",
+    "@react-navigation/native-stack": "^6.1.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-airbnb-base": "^14.2.1",
     "react": "17.0.1",
     "react-dom": "^17.0.2",
     "react-native": "0.64.1",
+    "react-native-safe-area-context": "^3.3.0",
+    "react-native-screens": "^3.5.0",
     "react-native-web": "^0.17.1"
   },
   "devDependencies": {
@@ -29,7 +33,7 @@
     "@types/jest": "^26.0.23",
     "@types/react-dom": "^17.0.9",
     "@types/react-native": "^0.64.5",
-    "@types/react-test-renderer": "^16.9.2",
+    "@types/react-test-renderer": "^16.9.5",
     "@typescript-eslint/eslint-plugin": "^4.29.1",
     "@typescript-eslint/parser": "^4.29.1",
     "babel-jest": "^26.6.3",
